Fix: TypeError in subprocess run of job_submitter.py
Previous (TypeError: __init__() got an unexpected keyword argument 'text') because text=True argument in the subprocess.run() function.

The text argument is available in Python 3.7 and later, but if you're using an older version of Python, you'll need to replace text=True with universal_newline=True, which works in both older and newer versions of Python.