refactor(ner): refactor ner_main use class variables instead multiple config reads + extract utility functions. Deprecate NER[article_limit] -> bathc_span_start/end.
Move away from inflexible and unclear article_limit config option which was actually a batch file span limiter.
New separate start and end more clearly deliniates not both are needed to be set. For example processing all batches beggining at X, which does not need an en index.

This is as part of a larger plan to make the ner module implementation completely independant on the config except for top level pipeline module initilization. This should provide far more robust and adaptable code with added benefit:

- self documenting which variables are actually used
- better refactoring
- IDE and type hinting